need to fixe time

+++ philo/src/parsing/init_struct.c
@@ -6,7 +6,7 @@
 /*   By: gaducurt <gaducurt@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/12 17:18:55 by gaducurt          #+#    #+#             */
-/*   Updated: 2025/10/01 14:17:13 by gaducurt         ###   ########.fr       */
+/*   Updated: 2025/10/02 15:15:22 by gaducurt         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,9 +20,9 @@ t_common	init_common(int ac, char **av)
 	t_common	common;
 
 	common.nb_philo = ft_atoi(av[1]);
-	common.time_to_die = ft_atoi(av[2]);
-	common.time_to_eat = ft_atoi(av[3]);
-	common.time_to_sleep = ft_atoi(av[4]);
+	common.time_to_die = (size_t)ft_atoi(av[2]);
+	common.time_to_eat = (size_t)ft_atoi(av[3]);
+	common.time_to_sleep = (size_t)ft_atoi(av[4]);
 	common.stop.data = 0;
 	if (ac == 6)
 		common.nb_must_eat = ft_atoi(av[5]);
